requirements.txt
webvtt-py
openai-whisper
moviepy
pytubefix
youtube-transcript-api
datasets